refactor: centralize sender admin metadata check
