+++ utils/adminUtils.js
@@ -25,6 +25,30 @@ function getEnvAdminSet() {
     .filter(Boolean));
 }
 
+/**
+ * Determine if the sender metadata on a message indicates admin privileges.
+ * @param {object} message - Message object containing sender metadata
+ * @returns {boolean} True if the sender appears to be an admin
+ */
+function senderIsAdminFromMessage(message) {
+  const sender = message?.sender;
+  if (!sender) return false;
+
+  if (sender.isAdmin === true || sender.isSuperAdmin === true) return true;
+
+  if (typeof sender.admin === 'string') {
+    const lowered = sender.admin.toLowerCase();
+    if (lowered.includes('admin')) return true;
+  }
+
+  if (Array.isArray(sender.labels)) {
+    return sender.labels.some(label => typeof label === 'string' && label.toLowerCase().includes('admin'));
+  }
+
+  return false;
+}
+
 module.exports = {
-  getEnvAdminSet
+  getEnvAdminSet,
+  senderIsAdminFromMessage
 };
